requirements.txt
streamlit
Pillow
requests
streamlit_lottie